% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahrd_on_gene_clusters.R
\name{annotateCluster}
\alias{annotateCluster}
\title{Each gene family (cluster of amino acid sequences) has members that in turn
might have InterPro annotations in argument 'ipr.annos'. Those are filtered
in order to retain only those that are NOT contained in others and those
that are NOT parents of others. Here, 'others' refers to InterPro
annotations found for genes of the cluster of course. The family 'fam' is
then assigned the most frequent InterPro family(s), if their respective
frequency is at least 0.5. If no such 'frequent' InterPro family annotation
is found, the most frequent InterPro annotation regardless of its type are
used. In most cases these are of type domain.}
\usage{
annotateCluster(fam, ipr.annos, interpro.database)
}
\arguments{
\item{fam}{a character vector of gene accessions}

\item{ipr.anno}{a data.frame with - at least - two columns, the first
holding the gene accessions, and the second holding the annotated InterPro
entry}

\item{interpro.databas}{the database of InterPro entries as parsed from the
interpro XML document 'interpro.xml'. The format is a named list of named
lists. See parseInterProXML(...) for more details.}
}
\value{
A list of InterPro accessions to be interpreted as the argument gene
family's 'fam' annotation. The list includes the annotations' frequency and
short descriptions.
}
\description{
Each gene family (cluster of amino acid sequences) has members that in turn
might have InterPro annotations in argument 'ipr.annos'. Those are filtered
in order to retain only those that are NOT contained in others and those
that are NOT parents of others. Here, 'others' refers to InterPro
annotations found for genes of the cluster of course. The family 'fam' is
then assigned the most frequent InterPro family(s), if their respective
frequency is at least 0.5. If no such 'frequent' InterPro family annotation
is found, the most frequent InterPro annotation regardless of its type are
used. In most cases these are of type domain.
}

