% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_description_funks.R
\name{annotateClusterWithProteinDescription}
\alias{annotateClusterWithProteinDescription}
\title{Generates a human readable description (HRD) for the argument
\code{prot.fam} based on the protein descriptions found for the genes
contained within the family. Candidate descriptions are filtered using
\code{AHRD.on.gene.clusters::filterTokens} and blacklisted with
\code{AHRD.on.gene.clusters::blacklistProtDescs}. The most frequent of the
retained descriptions are assigned as the clusters HRD. Maximum frequency
and Shannon Entropy are also returned as a measure of confidence.}
\usage{
annotateClusterWithProteinDescription(prot.fam, prot.desc.db,
  prot.id.col = "SHORT.ID", prot.desc.col = "filtered.description")
}
\arguments{
\item{prot.fam}{A character vector of gene accessions (IDs) forming the
argument protein family to annotate.}

\item{prot.desc.db}{A data.frame with protein descriptions for the family's
members. Typically the result of invoking
\code{AHRD.on.gene.clusters::readProteinDescriptionDb}.}

\item{prot.id.col}{A string or integer identifying the column of
\code{prot.desc.db} in which to find the protein IDs. Default is
\code{"SHORT.ID"}.}

\item{prot.desc.col}{A string or integer identifying the column of
\code{prot.desc.db} in which to find the protein descriptions. Default is
\code{"filtered.description"}.}
}
\value{
A list with the following key-value pairs: \code{descriptions} a
character vector of the most frequent filtered and blacklisted protein
descriptions, \code{max.frequency} the maximum frequency used to identify
the former protein descriptions, \code{entropy} the Shannon Entropy of the
filtered and blacklisted protein descriptions found for this protein family
\code{prot.fam}. \code{NA} is returned if no protein descriptions are
retained after filtering and blacklisting.
}
\description{
Generates a human readable description (HRD) for the argument
\code{prot.fam} based on the protein descriptions found for the genes
contained within the family. Candidate descriptions are filtered using
\code{AHRD.on.gene.clusters::filterTokens} and blacklisted with
\code{AHRD.on.gene.clusters::blacklistProtDescs}. The most frequent of the
retained descriptions are assigned as the clusters HRD. Maximum frequency
and Shannon Entropy are also returned as a measure of confidence.
}

