% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_description_funks.R
\name{blacklist}
\alias{blacklist}
\title{Function returns all those elements of \code{strs} that do not match any
regular expression in \code{blacklist.regexs}.}
\usage{
blacklist(strs, blacklist.regexs, ...)
}
\arguments{
\item{strs}{A character vector of strings to be submitted to blacklisting.}

\item{blacklist.regexs}{A character vector of regular expressions to be used
in blacklisting.}

\item{...}{optional additional arguments passed to \code{base::grepl}.}
}
\value{
A subset of \code{strs} of which no element matches any regular
expression in argument \code{blacklist.regexs}.
}
\description{
Function returns all those elements of \code{strs} that do not match any
regular expression in \code{blacklist.regexs}.
}

