% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_description_funks.R
\name{blacklistProtDescs}
\alias{blacklistProtDescs}
\title{Retains those argument \code{prot.descs} that do not match any regular
expression in the argument \code{blacklist.regexs}.}
\usage{
blacklistProtDescs(prot.descs,
  blacklist.regexs = getOption("AHRD.prot.desc.blacklist",
  readLines(file.path(path.package("AhrdOnGeneClusters"),
  "blacklist_descline.txt"))), ...)
}
\arguments{
\item{prot.descs}{A character vector holding protein descriptions to be
submitted to blacklisting.}

\item{blacklist.regexs}{A character vector of regular expression to be used
in the process of blacklisting. Default is
\code{getOption('AHRD.prot.desc.blacklist',
readLines(file.path(path.package('AhrdOnGeneClusters'),
'blacklist_descline.txt')))}}

\item{...}{optional additional parameters passed to function
\code{AhrdOnGeneClusters::blacklist}}
}
\value{
A subset of \code{prot.descs} of which no element matched any
regular expression in \code{blacklist.regexs}.
}
\description{
Retains those argument \code{prot.descs} that do not match any regular
expression in the argument \code{blacklist.regexs}.
}

