% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_description_funks.R
\name{blacklistTokens}
\alias{blacklistTokens}
\title{Retains those argument \code{tokens} that do not match any regular
expression in the argument \code{blacklist.regexs}.}
\usage{
blacklistTokens(tokens, blacklist.regexs = getOption("AHRD.token.blacklist",
  readLines(file.path(path.package("AHRD.on.gene.clusters"),
  "blacklist_token.txt"))), ...)
}
\arguments{
\item{tokens}{A character vector holding protein descriptions to be
submitted to blacklisting.}

\item{blacklist.regexs}{A character vector of regular expression to be used
in the process of blacklisting. Default is
\code{getOption('AHRD.on.gene.clusters',
readLines(file.path(path.package('AHRD.on.gene.clusters'),
'blacklist_token.txt')))}}

\item{...}{optional additional parameters passed to function
\code{AHRD.on.gene.clusters::blacklist}}
}
\value{
A subset of \code{tokens} of which no element matched any regular
expression in \code{blacklist.regexs}.
}
\description{
Retains those argument \code{tokens} that do not match any regular
expression in the argument \code{blacklist.regexs}.
}

