% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_description_funks.R
\name{filter}
\alias{filter}
\title{Within each element of argument \code{strs} all substrings matching
any of the regular expressions in argument \code{filter.regexs} is deleted.
Finally leading and trailing white spaces are also abolished.}
\usage{
filter(strs, filter.regexs, ...)
}
\arguments{
\item{strs}{A character vector of protein descriptions to be filtered.}

\item{filter.regexs}{A character vector of regular expressions to be used
within filtering.}

\item{...}{optional additional parameters passed to function
\code{base::gsub}.}
}
\value{
A modified version of \code{strs} in which all matches to any
regular expression in \code{filter.regexs}, along with leading or trailing
white spaces has been deleted.
}
\description{
Within each element of argument \code{strs} all substrings matching
any of the regular expressions in argument \code{filter.regexs} is deleted.
Finally leading and trailing white spaces are also abolished.
}

