% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahrd_on_gene_clusters.R
\name{filterInterProClusterAnnotations}
\alias{filterInterProClusterAnnotations}
\title{Reduce a set of InterPro accessions (IDs) to those not being contained by
others, and those not being parents of others. Hence give the 'most
informative' InterPro accessions of the argument 'ipr.ids'.}
\usage{
filterInterProClusterAnnotations(ipr.ids, interpro.database = ipr.db)
}
\arguments{
\item{ipr.id}{Result of parsing the InterPro XML database}
}
\value{
The character vector of InterPro accessions, that is the subset of
argument 'irp.ids' filtered as descibed above.
}
\description{
Reduce a set of InterPro accessions (IDs) to those not being contained by
others, and those not being parents of others. Hence give the 'most
informative' InterPro accessions of the argument 'ipr.ids'.
}

