% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_description_funks.R
\name{filterProtDescs}
\alias{filterProtDescs}
\title{Within each element of argument \code{prot.descs} all substrings matching
any of the regular expressions in argument \code{filter.regexs} is deleted.
Finally leading and trailing white spaces are also abolished.}
\usage{
filterProtDescs(prot.descs, filter.regexs = getOption("AHRD.prot.desc.filter",
  readLines(file.path(path.package("AHRD.on.gene.clusters"),
  "filter_descline.txt"))), ...)
}
\arguments{
\item{prot.descs}{A character vector of protein descriptions to be filtered.}

\item{filter.regexs}{A character vector of regular expressions to be used
within filtering. Default is \code{getOption('AHRD.prot.desc.filter',
readLines(file.path(path.package('AHRD.on.gene.clusters'),
'filter_descline.txt')))}}

\item{...}{optional additional parameters passed to function
\code{AHRD.on.gene.clusters::filter}.}
}
\value{
A modified version of \code{prot.descs} in which all matches to any
regular expression in \code{filter.regexs}, along with leading or trailing
white spaces has been deleted.
}
\description{
Within each element of argument \code{prot.descs} all substrings matching
any of the regular expressions in argument \code{filter.regexs} is deleted.
Finally leading and trailing white spaces are also abolished.
}

