% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahrd_on_gene_clusters.R
\name{interProRecursion}
\alias{interProRecursion}
\title{Recursively traverses InterPro entries held in argument 'interpro.database'
and finds all entries related to argument 'ipr.id' through relation
'relship.slot'. In case of PARENTS all parents and parents of parents are
returned, in case of CONTAINS all contained and contained in contained are.}
\usage{
interProRecursion(ipr.id, relship.slot = "PARENTS",
  interpro.database = ipr.db)
}
\arguments{
\item{ipr.i}{he InterPro accession to start the recursion from, for example
'IPR000003'}

\item{relship.slo}{Either 'PARENTS' or 'CONTAINS', default is 'PARENTS'}

\item{interpro.databas}{The named list containing the parsed contents of
the XML interpro document. See parseInterProXML(...) for details.}
}
\value{
A character vector of recursive PARENTS or CONTAINED InterPro
entries for the argument entry 'ipr.id'.
}
\description{
Recursively traverses InterPro entries held in argument 'interpro.database'
and finds all entries related to argument 'ipr.id' through relation
'relship.slot'. In case of PARENTS all parents and parents of parents are
returned, in case of CONTAINS all contained and contained in contained are.
}

