% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahrd_on_gene_clusters.R
\name{parseInterProTag}
\alias{parseInterProTag}
\title{Use library XML to parse a single InterPro tag and extract ID, SHORT.NAME,
PARENT-, and CONTAINS-list.}
\usage{
parseInterProTag(xml.node, include.abstracts = TRUE)
}
\arguments{
\item{xml.node}{A single entry of the result of calling getNodeSet(...) on
the InterPro XML document}
}
\value{
A list with three attributes 'ID', 'PARENTS', and 'CONTAINS'.
}
\description{
Use library XML to parse a single InterPro tag and extract ID, SHORT.NAME,
PARENT-, and CONTAINS-list.
}

