% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{testAnnotateCluster}
\alias{testAnnotateCluster}
\title{Function tests \code{annotateCluster} using real life data stored in this'
package's data \code{ipr_and_families.RData}.}
\usage{
testAnnotateCluster()
}
\value{
TRUE if and only if no error has occurred.
}
\description{
Function tests \code{annotateCluster} using real life data stored in this'
package's data \code{ipr_and_families.RData}.
}

