% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahrd_on_gene_clusters.R
\name{parseInterProXML}
\alias{parseInterProXML}
\title{Read and parse the InterPro XML database.}
\usage{
parseInterProXML(path.2.xml, include.abstracts = TRUE)
}
\arguments{
\item{path.2.xml}{The file path to the interpro.xml document}

\item{include.abstracts}{If TRUE the quite long abstract texts will be
added to the result.}
}
\value{
A named list of lists, each of which is the result of parsing a
single InterPro XML node with function parseInterProTag(...).
}
\description{
Read and parse the InterPro XML database.
}

